@@ -1,6 +1,6 @@
 package constant
 
 const (
-	GSHEET_CREDENTIAL = "GSHEET_CREDENTIAL" // #nosec G101
-	GSHEET_TOKEN      = "GSHEET_TOKEN"
+	GsheetCredential = "GSHEET_CREDENTIAL" // #nosec G101
+	GsheetToken      = "GSHEET_TOKEN"
 )
